[cli] bench: `pocket-ic` replica
